[xh:oc] use only character trie to search syllable for later lattice building

+++ src/ime-core/view/imi_view_classic.cpp
@@ -67,7 +67,7 @@ CIMIClassicView::updateWindows(unsigned mask)
             if (word_num == 0) goto pass;  // when sentence is not worthy of
 
 #ifdef DEBUG
-            printf("%d ", i);
+            printf("%d ", (int)i);
             print_wide(sentence.c_str());
             printf("\n");
 #endif
